tests/requirements.txt
pytest==8.3.4
requests==2.32.3
python-dotenv==1.0.1

